+++ test/template.js
@@ -293,6 +293,21 @@ describe('Template class tests', function () {
                 );
             });
     });
+    it('schema_one_of', function () {
+        const ymldata = fs.readFileSync(`${templatesPath}/combine.yml`, 'utf8');
+        return Template.loadYaml(ymldata)
+            .then((tmpl) => {
+                const schema = tmpl.getParametersSchema();
+                console.log(JSON.stringify(schema, null, 2));
+
+                assert.ok(schema.properties.selection);
+                const selection = schema.properties.selection;
+                assert.ok(selection.oneOf);
+                assert.ok(selection.oneOf[0].properties.prop1);
+                assert.ok(selection.oneOf[1].properties.prop1);
+                assert.ok(selection.oneOf[1].properties.prop2);
+            });
+    });
     it('render', function () {
         const mstdata = `
             {{foo::string}}